requirements.txt
playwright==1.41.1
pydantic==2.5.3
aiohttp==3.9.1
pandas==2.1.4
python-dotenv==1.0.0
PyYAML==6.0.1
openai==1.6.1
python-json-logger==2.0.7
aiohttp-socks==0.8.4
tenacity==8.2.3